Allow for partial config files
This change allows Yeogurt to read incomplete .yo-rc.json files. If only some of the options are specified and the user selects 'Yes' to the "Use existing config" prompt, then Yeogurt will skip the steps that are already specified in the partial .yo-rc.json and prompt for the steps that are missing.

This is useful for generator composability, such as when an external generator includes Yeogurt and may pass it some predetermined setup options, but leave others up to the user to decide.